event-sector: test containers configuration
