+++ src/test/kotlin/com/ticketera/repositories/TicketTypeRepositoryTest.kt
@@ -11,7 +11,6 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest
 import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager
 import org.springframework.test.context.DynamicPropertyRegistry
 import org.springframework.test.context.DynamicPropertySource
-import org.testcontainers.containers.PostgreSQLContainer
 import org.testcontainers.junit.jupiter.Container
 import org.testcontainers.junit.jupiter.Testcontainers
 
@@ -61,23 +60,11 @@ final class TicketTypeRepositoryTest : TestData() {
 
     companion object {
         @Container
-        val postgres = PostgreSQLContainer<Nothing>("postgres:16.4")
-            .apply {
-                withDatabaseName("ticketera")
-                withUsername("ticketera")
-                withPassword("dGVzdGRi")
-                withReuse(true)
-            }
+        val postgres = TestContainerConf.postgres
 
         @JvmStatic
         @DynamicPropertySource
-        fun registerProperties(registry: DynamicPropertyRegistry) {
-            registry.add("spring.datasource.url", postgres::getJdbcUrl)
-            registry.add("spring.datasource.username", postgres::getUsername)
-            registry.add("spring.datasource.password", postgres::getPassword)
-            registry.add("spring.flyway.url", postgres::getJdbcUrl)
-            registry.add("spring.flyway.user", postgres::getUsername)
-            registry.add("spring.flyway.password", postgres::getPassword)
-        }
+        fun registerProperties(registry: DynamicPropertyRegistry) =
+            TestContainerConf.registerProperties(registry)
     }
 }
